requirement.txt
alembic==1.4.1
aniso8601==8.0.0
apturl==0.5.2
asn1crypto==0.24.0
blinker==1.4
Brlapi==0.6.6
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
colorama==0.3.7
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
dnspython==1.16.0
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Markdown==0.3
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-PageDown==0.2.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
google-api-python-client==1.7.11
google-auth==1.10.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
gunicorn==20.0.4
httplib2==0.16.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.1.2
Markdown==3.1.1
MarkupSafe==1.1.1
mysql-connector-python==8.0.19
netifaces==0.10.4
oauth==1.0.1
oauthlib==3.1.0
olefile==0.45.1
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pyinotify==0.9.6
pymacaroons==0.13.0
PyNaCl==1.1.2
pyOpenSSL==17.5.0
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.8.1
python-debian==0.1.32
python-editor==1.0.4
pytz==2019.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
rsa==4.0
SecretStorage==2.3.1
simplejson==3.13.2
six==1.14.0
SQLAlchemy==1.3.15
system-service==0.3
systemd-python==234
tzlocal==2.0.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.25.8
usb-creator==0.3.3
virtualenv==16.7.7
visitor==0.1.3
wadllib==1.3.2
Werkzeug==0.16.1
WTForms==2.2.1
xkit==0.0.0
zope.interface==4.3.2
